邮件抓取下载功能实现

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/drawback/grabfile/form/TGrabFileAddForm.java
@@ -12,13 +12,13 @@ import com.uxuexi.core.web.form.AddForm;
 public class TGrabFileAddForm extends AddForm {
 
 	/**主键*/
-	private Integer id;
+	private long id;
 
 	/**抓取邮件id*/
-	private Integer mailId;
+	private long mailId;
 
 	/**上级id*/
-	private Integer parentId;
+	private long parentId;
 
 	/**文件夹名称*/
 	private String folderName;
@@ -33,10 +33,10 @@ public class TGrabFileAddForm extends AddForm {
 	private String fileSize;
 
 	/**文件类型(1-文件夹;2-文件;)*/
-	private Integer type;
+	private long type;
 
 	/**状态(1-已删除;2-已启用;)*/
-	private Integer status;
+	private long status;
 
 	/**创建时间*/
 	private Date createTime;
@@ -45,15 +45,15 @@ public class TGrabFileAddForm extends AddForm {
 	private Date updateTime;
 
 	/**层级*/
-	private Integer level;
+	private long level;
 
 	/**完整路径*/
 	private String fullPath;
 
 	/**序号*/
-	private Integer sort;
+	private long sort;
 
 	/**序号*/
-	private Integer groupType;
+	private long groupType;
 
 }
